# Conflicts:
#	server/src/main/java/mainproject33/domain/comment/controller/CommentController.java
#	server/src/main/java/mainproject33/domain/comment/entity/Comment.java
#	server/src/main/java/mainproject33/domain/comment/service/CommentService.java
#	server/src/main/java/mainproject33/domain/like/controller/LikeController.java
#	server/src/main/java/mainproject33/domain/like/entity/Like.java
#	server/src/main/java/mainproject33/domain/like/repository/LikeRepository.java
#	server/src/main/java/mainproject33/domain/like/service/LikeService.java
#	server/src/main/java/mainproject33/domain/userboard/entity/UserBoard.java
#	server/src/main/java/mainproject33/domain/userboard/mapper/UserBoardMapper.java

+++ server/src/main/java/mainproject33/domain/like/service/LikeService.java
@@ -3,11 +3,11 @@ package mainproject33.domain.like.service;
 import lombok.RequiredArgsConstructor;
 import mainproject33.domain.comment.entity.Comment;
 import mainproject33.domain.comment.service.CommentService;
+import mainproject33.domain.like.entity.Like;
+import mainproject33.domain.like.repository.LikeRepository;
 import mainproject33.domain.member.entity.Member;
 import mainproject33.domain.member.service.MemberService;
 import mainproject33.domain.userboard.entity.UserBoard;
-import mainproject33.domain.like.entity.Like;
-import mainproject33.domain.like.repository.LikeRepository;
 import mainproject33.domain.userboard.service.UserBoardService;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -25,49 +25,43 @@ public class LikeService
 
     private final CommentService commentService;
 
-    public Like upLikeBoard(long memberId, long boardId)
+    public boolean changeBoardLike(Member member, long boardId)
     {
         UserBoard userBoard = boardService.findUserBoard(boardId);
-        Member member = memberService.findVerifiedMember(memberId);
 
-        Like boardLike = boardLikeRepository.save(new Like(member, userBoard));
-
-
-        return boardLike;
+        if(checkLikedBoard(member.getId(), boardId))
+        {
+            boardLikeRepository.save(new Like(member, userBoard));
+            return true;
+        }
+        else
+        {
+            userBoard.unLike();
+            boardLikeRepository.deleteByMemberAndUserBoard(member, userBoard);
+            return false;
+        }
     }
 
-    public void unlikeBoard(long memberId, long boardId)
+    public boolean changeCommentLike(Member member, long commentId)
     {
-        UserBoard userBoard = boardService.findUserBoard(boardId);
-        Member member = memberService.findVerifiedMember(memberId);
-
-        userBoard.unLike();
-        boardLikeRepository.deleteByMemberAndUserBoard(member, userBoard);
-    }
-
-    public Like upLikeComment(long memberId, long commentId)
-    {
-        Member member = memberService.findVerifiedMember(memberId);
         Comment comment = commentService.findComment(commentId);
 
-        return boardLikeRepository.save(new Like(member, comment));
+        if(checkLikedComment(member.getId(), comment.getId()))
+        {
+            boardLikeRepository.save(new Like(member, comment));
+            return true;
+        }
+
+        else
+        {
+            comment.unlike();
+            boardLikeRepository.deleteByMemberAndComment(member, comment);
+            return false;
+        }
     }
 
-    public void unlikeComment(long memberId, long commentId)
-    {
-        Member member = memberService.findVerifiedMember(memberId);
-        Comment comment = commentService.findComment(commentId);
-
-        comment.unlike();
-        boardLikeRepository.deleteByMemberAndComment(member, comment);
-    }
-
-
-
-    /**
-     * 좋아요 했느지 여부 검사
-     */
-    public boolean checkLikedComment(long memberId, long commentId)
+    //좋아요 했는지 여부 검사
+    private boolean checkLikedComment(long memberId, long commentId)
     {
         Member member = memberService.findVerifiedMember(memberId);
         Comment comment = commentService.findComment(commentId);
@@ -75,11 +69,13 @@ public class LikeService
         return boardLikeRepository.findByMemberAndComment(member, comment)
                 .isEmpty();
     }
-    public boolean checkLikedBoard(long memberId, long boardId)
+
+    private boolean checkLikedBoard(long memberId, long boardId)
     {
         Member member = memberService.findVerifiedMember(memberId);
         UserBoard board = boardService.findUserBoard(boardId);
 
+        //비어 있으면 좋아요를 안한 것 true 리턴
         return boardLikeRepository.findByMemberAndUserBoard(member, board)
                 .isEmpty();
     }
